+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/borders_flow_controller.h
@@ -42,7 +42,7 @@ public:
     void OnReadyMergeBorders();
 
     void Enqueue(const TEvBordersConstructionResult::TPtr& event);
-    
+
     ~TBordersFlowController();
 
 private:
