+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/events.cpp
@@ -15,4 +15,21 @@ TEvRequestFilter::TEvRequestFilter(const TPortionDataSource& source, const std::
 {
 }
 
+TEvRequestFilter::TEvRequestFilter(const NArrow::TSimpleRow& minPK, const NArrow::TSimpleRow& maxPK, const ui64 portionId,
+    const ui64 recordsCount, const TSnapshot& maxVersion, const std::shared_ptr<IFilterSubscriber>& subscriber,
+    const std::shared_ptr<const TAtomicCounter>& abortionFlag)
+    : MinPK(minPK)
+    , MaxPK(maxPK)
+    , PortionId(portionId)
+    , RecordsCount(recordsCount)
+    , MaxVersion(maxVersion)
+    , Subscriber(subscriber)
+    , AbortionFlag(abortionFlag)
+{
+}
+
+TSnapshot TEvRequestFilter::GetMaxVersion() const {
+    return MaxVersion;
+}
+
 }   // namespace NKikimr::NOlap::NReader::NSimple::NDuplicateFiltering
