+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/private_events.h
@@ -18,21 +18,10 @@ private:
 
 public:
     TEvFilterRequestResourcesAllocated(const std::shared_ptr<TFilterAccumulator>& request,
-        const std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>& guard, std::unique_ptr<TFilterBuildingGuard>&& requestGuard)
-        : Request(request)
-        , AllocationGuard(guard)
-        , RequestGuard(std::move(requestGuard))
-    {
-        AFL_VERIFY(RequestGuard);
-    }
+        const std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>& guard, std::unique_ptr<TFilterBuildingGuard>&& requestGuard);
 
-    std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>&& ExtractAllocationGuard() {
-        return std::move(AllocationGuard);
-    }
-    std::unique_ptr<TFilterBuildingGuard>&& ExtractRequestGuard() {
-        AFL_VERIFY(RequestGuard);
-        return std::move(RequestGuard);
-    }
+    std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>&& ExtractAllocationGuard();
+    std::unique_ptr<TFilterBuildingGuard>&& ExtractRequestGuard();
 };
 
 }   // namespace NKikimr::NOlap::NReader::NSimple::NDuplicateFiltering::NPrivate
@@ -48,20 +37,10 @@ private:
 
 public:
     TBuildFilterTaskContext(
-        TBuildFilterContext&& context, const std::shared_ptr<TBuildFilterTaskExecutor>& executor, TBordersBatch&& batch)
-        : Context(std::move(context))
-        , Executor(executor)
-        , Batch(std::move(batch))
-    {
-    }
+        TBuildFilterContext&& context, const std::shared_ptr<TBuildFilterTaskExecutor>& executor, TBordersBatch&& batch);
 
-    const TBuildFilterContext& GetGlobalContext() const {
-        return Context;
-    }
-
-    TBuildFilterContext&& ExtractGlobalContext() {
-        return std::move(Context);
-    }
+    const TBuildFilterContext& GetGlobalContext() const;
+    TBuildFilterContext&& ExtractGlobalContext();
 };
 
 class TDuplicateSourceCacheResult {
@@ -70,38 +49,10 @@ private:
     TColumnData DataByAddress;
 
 public:
-    TDuplicateSourceCacheResult(TColumnData&& data)
-        : DataByAddress(std::move(data))
-    {
-    }
+    TDuplicateSourceCacheResult(TColumnData&& data);
 
     THashMap<ui64, std::shared_ptr<NArrow::TGeneralContainer>> ExtractDataByPortion(
-        const std::map<ui32, std::shared_ptr<arrow::Field>>& fieldByColumn) {
-        THashMap<ui64, std::shared_ptr<NArrow::TGeneralContainer>> dataByPortion;
-        std::vector<std::shared_ptr<arrow::Field>> fields;
-        for (const auto& [_, field] : fieldByColumn) {
-            fields.emplace_back(field);
-        }
-
-        THashMap<ui64, THashMap<ui32, std::shared_ptr<NArrow::NAccessor::IChunkedArray>>> columnsByPortion;
-        for (auto&& [address, data] : DataByAddress) {
-            AFL_VERIFY(columnsByPortion[address.GetPortionId()].emplace(address.GetColumnId(), data).second);
-        }
-
-        for (auto& [portion, columns] : columnsByPortion) {
-            std::vector<std::shared_ptr<NArrow::NAccessor::IChunkedArray>> sortedColumns;
-            for (const auto& [columnId, _] : fieldByColumn) {
-                auto column = columns.FindPtr(columnId);
-                AFL_VERIFY(column);
-                sortedColumns.emplace_back(*column);
-            }
-            std::shared_ptr<NArrow::TGeneralContainer> container =
-                std::make_shared<NArrow::TGeneralContainer>(fields, std::move(sortedColumns));
-            AFL_VERIFY(dataByPortion.emplace(portion, std::move(container)).second);
-        }
-
-        return dataByPortion;
-    }
+        const std::map<ui32, std::shared_ptr<arrow::Field>>& fieldByColumn);
 };
 
 class TEvBordersConstructionResult
@@ -114,17 +65,10 @@ public:
 public:
     TEvBordersConstructionResult(TBuildFilterTaskContext&& context,
         THashMap<NGeneralCache::TGlobalColumnAddress, std::shared_ptr<NArrow::NAccessor::IChunkedArray>>&& columns,
-        const std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>& allocationGuard)
-        : Context(std::move(context))
-        , Result(std::move(columns))
-        , AllocationGuard(allocationGuard)
-    {}
+        const std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>& allocationGuard);
 
     TEvBordersConstructionResult(TBuildFilterTaskContext&& context,
-        TConclusion<TDuplicateSourceCacheResult>&& error)
-        : Context(std::move(context))
-        , Result(std::move(error))
-    {}
+        TConclusion<TDuplicateSourceCacheResult>&& error);
 };
 
 class TEvMergeBordersResult
@@ -135,11 +79,7 @@ public:
     TConclusionStatus Result;
 
 public:
-    TEvMergeBordersResult(TBuildFilterTaskContext&& context, THashMap<ui64, NArrow::TColumnFilter>&& readyFilters, TConclusionStatus&& conclusion)
-        : Context(std::move(context))
-        , ReadyFilters(std::move(readyFilters))
-        , Result(std::move(conclusion)) {
-    }
+    TEvMergeBordersResult(TBuildFilterTaskContext&& context, THashMap<ui64, NArrow::TColumnFilter>&& readyFilters, TConclusionStatus&& conclusion);
 };
 
 }
